requirements.txt
bertopic
chainlit
dash.ly
langchain
langchain-chroma
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
nltk
#numpy
openai
pandas
plotly
PyMuPDF
pymupdf4llm
python-dotenv
pyyaml
scipy
sentence-transformers
scikit-learn
spacy
